+++ packages/okam-build/test/tasks/processor/template/transform/base/style.spec.js
@@ -37,6 +37,15 @@ describe('style binding', function () {
             '<view style="height:10px;color:{{colorStyle}};font-size:{{fontStyle + \'px\'}}"></view>');
     });
 
+    it('should support style binding syntax which contains ternary expression', function () {
+        const file = {
+            content: '<div :style="{height: flag === \'true\' || flag === true ? height + \'px\' : \'20px\', width: flag ? \'10px\' : \'30px\'}"></div>'
+        };
+        const result = templateProcessor(file, fakeProcessorOptions());
+        assert.equal(result.content,
+            '<view style="height:{{flag === \'true\' || flag === true ? height + \'px\' : \'20px\'}};width:{{flag ? \'10px\' : \'30px\'}}"></view>');
+    });
+
     // 和对象语法走的同一个代码分支。转化的时候应避免内部[]被转化
     it('should transform style-data-binding with array that has object（数组语法）', function () {
         const file = {
@@ -45,5 +54,36 @@ describe('style binding', function () {
         const result = templateProcessor(file, fakeProcessorOptions());
         assert.equal(result.content,
             '<view style="color:{{colorStyle[item]}};font-size:{{fontStyle + \'px\'}};font-weight:{{\'bold\'}}"></view>');
+    });
+
+    it('should support object abbrev style binding syntax', function () {
+        const file = {
+            content: '<div :style="{height, fontSize}"></div>'
+        };
+        const result = templateProcessor(file, fakeProcessorOptions());
+        assert.equal(result.content,
+            '<view style="height:{{height}};font-size:{{fontSize}}"></view>');
+    });
+
+    it('should support transform style binding syntax with comma', function () {
+        const file = {
+            content: '<div :style="{ transform: \'translate(-50%, -50%) rotate(\' + preview.rotate + \'deg)\', fontSize, \'font-size\': \'12px\' }"></div>'
+        };
+        const result = templateProcessor(file, fakeProcessorOptions());
+        assert.equal(
+            result.content,
+            '<view style="transform:{{\'translate(-50%, -50%) rotate(\' + preview.rotate + \'deg)\'}};font-size:{{fontSize}};\'font-size\':{{\'12px\'}}"></view>'
+        );
+    });
+
+    it('should support es6 template string style binding', function () {
+        const file = {
+            content: '<div :style="{fontSize: `${fontSize}px`, width}"></div>'
+        };
+        const result = templateProcessor(file, fakeProcessorOptions());
+        assert.equal(
+            result.content,
+            '<view style="font-size:{{\'\' + fontSize + \'px\'}};width:{{width}}"></view>'
+        );
     });
 });